refactor: update component to use tabs instead of buttons
